moving lib/auth/kube to lib/auth/kubecluster
